Move some of the functions over to jest-preset
