+++ packages/testing/config/jest/api/jest.setup.js
@@ -1,126 +1,23 @@
 // /* eslint-env jest */
 // @ts-check
-const fs = require('fs')
-const path = require('path')
 
+// @NOTE without these imports in the setup file, mockCurrentUser
+// and defineScenario remain undefined in the user's tests
 const { setContext } = require('@redwoodjs/graphql-server/dist/globalContext')
-const { defineScenario } = require('@redwoodjs/testing/dist/api/scenario')
+const { defineScenario } = require('@redwoodjs/testing/dist/api')
 
 // @NOTE we do this because jest.setup.js runs every time in each worker
 // while jest-preset runs once. This significantly reduces memory footprint, and testing time
 // The key is to reduce the amount of imports in this file because of how jest handles require caching
-const { getDMMF, getPrismaConfig, rwPaths } = global.__RWJS__TEST_IMPORTS
+const { configureTeardown, teardown, disconnect, seedScenario } =
+  global.__RWJS__TEST_IMPORTS
 
-// Error codes thrown by [MySQL, SQLite, Postgres] when foreign key constraint
-// fails on DELETE
-const FOREIGN_KEY_ERRORS = [1451, 1811, 23503]
 const DEFAULT_SCENARIO = 'standard'
-const TEARDOWN_CACHE_PATH = path.join(
-  rwPaths.generated.base,
-  'scenarioTeardown.json'
-)
-let teardownOrder = []
-let originalTeardownOrder = []
 
-const deepCopy = (obj) => {
-  return JSON.parse(JSON.stringify(obj))
-}
-
-const isIdenticalArray = (a, b) => {
-  return JSON.stringify(a) === JSON.stringify(b)
-}
-
-const configureTeardown = async () => {
-  // @NOTE prisma utils are available in cli lib/schemaHelpers
-  // But avoid importing them, to prevent memory leaks in jest
-  const schema = await getDMMF({ datamodelPath: rwPaths.api.dbSchema })
-  const schemaModels = schema.datamodel.models.map((m) => m.dbName || m.name)
-
-  // check if pre-defined delete order already exists and if so, use it to start
-  if (fs.existsSync(TEARDOWN_CACHE_PATH)) {
-    teardownOrder = JSON.parse(fs.readFileSync(TEARDOWN_CACHE_PATH).toString())
-  }
-
-  // check the number of models in case we've added/removed since cache was built
-  if (teardownOrder.length !== schemaModels.length) {
-    teardownOrder = schemaModels
-  }
-
-  // keep a copy of the original order to compare against
-  originalTeardownOrder = deepCopy(teardownOrder)
-}
-
-const seedScenario = async (scenario) => {
-  console.log('in seed Scenario', scenario)
-  if (scenario) {
-    const { db } = require(path.join(rwPaths.api.src, 'lib', 'db'))
-
-    const scenarios = {}
-    for (const [model, namedFixtures] of Object.entries(scenario)) {
-      scenarios[model] = {}
-      for (const [name, createArgs] of Object.entries(namedFixtures)) {
-        if (typeof createArgs === 'function') {
-          scenarios[model][name] = await db[model].create(createArgs(scenarios))
-        } else {
-          scenarios[model][name] = await db[model].create(createArgs)
-        }
-      }
-    }
-    return scenarios
-  } else {
-    return {}
-  }
-}
-
-const teardown = async () => {
-  // Don't populate global scope, keep util functions inside teardown
-  // determine what kind of quotes are needed around table names in raw SQL
-  const getQuoteStyle = async () => {
-    const config = await getPrismaConfig({
-      datamodel: fs.readFileSync(rwPaths.api.dbSchema).toString(),
-    })
-
-    switch (config.datasources?.[0]?.provider) {
-      case 'mysql':
-        return '`'
-      default:
-        return '"'
-    }
-  }
-
-  const quoteStyle = await getQuoteStyle()
-
-  const { db } = require(path.join(rwPaths.api.src, 'lib', 'db'))
-
-  for (const modelName of teardownOrder) {
-    try {
-      await db.$executeRawUnsafe(
-        `DELETE FROM ${quoteStyle}${modelName}${quoteStyle}`
-      )
-    } catch (e) {
-      const match = e.message.match(/Code: `(\d+)`/)
-      if (match && FOREIGN_KEY_ERRORS.includes(parseInt(match[1]))) {
-        const index = teardownOrder.indexOf(modelName)
-        teardownOrder[index] = null
-        teardownOrder.push(modelName)
-      } else {
-        throw e
-      }
-    }
-  }
-
-  // remove nulls
-  teardownOrder = teardownOrder.filter((val) => val)
-
-  // if the order of delete changed, write out the cached file again
-  if (!isIdenticalArray(teardownOrder, originalTeardownOrder)) {
-    originalTeardownOrder = deepCopy(teardownOrder)
-    fs.writeFileSync(TEARDOWN_CACHE_PATH, JSON.stringify(teardownOrder))
-  }
-}
-
+// @NOTE without this function in jest-setup, secnario comes through as undefined
+// Not exactly clear why...
 const buildScenario =
-  (it) =>
+  (it, testPath) =>
   (...args) => {
     let scenarioName, testName, testFunc
 
@@ -135,7 +32,7 @@ const buildScenario =
 
     return it(testName, async () => {
       const path = require('path')
-      const testFileDir = path.parse(global.testPath)
+      const testFileDir = path.parse(testPath)
       // e.g. ['comments', 'test'] or ['signup', 'state', 'machine', 'test']
       const testFileNameParts = testFileDir.name.split('.')
       const testFilePath = `${testFileDir.dir}/${testFileNameParts
@@ -169,8 +66,8 @@ const buildScenario =
     })
   }
 
-global.scenario = buildScenario(global.it)
-global.scenario.only = buildScenario(global.it.only)
+global.scenario = buildScenario(global.it, global.testPath)
+global.scenario.only = buildScenario(global.it.only, global.testPath)
 
 global.defineScenario = defineScenario
 
@@ -185,8 +82,7 @@ beforeAll(async () => {
 })
 
 afterAll(async () => {
-  const { db } = require(path.join(rwPaths.api.src, 'lib', 'db'))
-  await db.$disconnect()
+  await disconnect()
 })
 
 afterEach(async () => {
